+++ srcs/ft_print_p.c
@@ -6,7 +6,7 @@
 /*   By: emaugale <emaugale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/31 19:42:46 by emaugale          #+#    #+#             */
-/*   Updated: 2021/09/18 00:18:45 by emaugale         ###   ########.fr       */
+/*   Updated: 2021/09/19 01:02:35 by emaugale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_print_p(va_list args)
 {
-	int	nbr;
+	unsigned long long	nbr;
 
-	nbr = va_arg(args, int);
+	nbr = va_arg(args, unsigned long long);
 	ft_putstr("0x");
 	ft_putnbr_base_unsigned(nbr, "0123456789abcdef");
 	return (1);
